Refactored the `logo` class name in 404.html
The `.logo` class in the main CSS file applies a transparent
image (spacer) using a relative path. However, this 404.html
file is not managed in Flare, so the relative paths do not
work. Changed this class to `.logo-for-404` in the file header,
and added styles for width and height as per the class in the
main CSS file.
Also, reverted the commented-out `.logo-wrapper` div.