Adding XML files to script parsing
The sitemaps (EN and JA) are in XML format. Each publication has
four or five XML files total.
We want to apply our string replacement script to the JA sitemaps.
This change adds a check for XML file extension, enabling parsing
of these files.
Our string replacement script already includes pairs for the
publication base links. We're applying that to the sitemaps.