Updating tests
We refactored the build output to include the 'Content/index.htm' file.
The EN file hides the left-side nav, while the JA file presently
does not have this structure. Adding a test for this difference.
The JA file will have the same structure as EN file following next
translation run.